Refactoring test cases
